expression api fixed

+++ src/main/java/com/PartnersFunds/Repo/ExpressionConditionRepo.java
@@ -3,5 +3,5 @@ package com.PartnersFunds.Repo;
 import com.PartnersFunds.Entities.ExpressionConditionEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ExpressionConditionRepo extends JpaRepository<ExpressionConditionEntity, Long> {
+public interface ExpressionConditionRepo extends JpaRepository<ExpressionConditionEntity, Integer> {
 }
