+++ src/main/java/com/PartnersFunds/Repo/ExpressionRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface ExpressionRepo extends JpaRepository<ExpressionEntity, Long> {
+public interface ExpressionRepo extends JpaRepository<ExpressionEntity, Integer> {
 	
 	@Query(value="select EXPRESSION_NAME from XXPF_EXPRESSIONS where EXPRESSION_ID = :exp_id",nativeQuery=true)
 	String ExpNameQuery(@Param("exp_id") int exp_id);
